use StandardButton directly in components
